--- okhttp-ktx/Module.md
@@ -1,3 +1,0 @@
-# Module okhttp-ktx
-
-OkHttp KTX library.
