--- okhttp-ktx/README.md
@@ -1,4 +1,0 @@
-OkHttp KTX
-==========
-
-Komfortable APIs for Kotlin clients.
