--- okhttp-ktx/api/okhttp-ktx.api
@@ -1,5 +1,0 @@
-public final class okhttp3/JvmCallExtensionsKt {
-	public static final fun awaitResponse (Lokhttp3/Call;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public static final fun executeAsync (Lokhttp3/Call;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
